Resolve TokenGrantingRelayer
