--- universal-login-relayer/knexfile.js
@@ -1,1 +1,0 @@
-module.exports = require('./lib/config/knexfile.js')
