+++ universal-login-relayer/knexfile.ts
@@ -1,0 +1,3 @@
+import knexfile from './lib/config/knexfile.js';
+
+export default knexfile;
