--- universal-login-relayer/lib/dev/TokenGrantingRelayer.js
@@ -1,38 +1,0 @@
-import {waitToBeMined} from '@universal-login/commons';
-import Token from './Token.json';
-import Relayer from '@universal-login/relayer';
-import {utils, Contract} from 'ethers';
-
-class TokenGrantingRelayer extends Relayer {
-  constructor(config, provider = '') {
-    super(config, provider);
-    this.tokenContractAddress = config.tokenContractAddress;
-    this.tokenContract = new Contract(this.tokenContractAddress, Token.interface, this.wallet);
-    this.addHooks();
-  }
-
-  addHooks() {
-    this.hooks.addListener('created', async (transaction) => {
-      const receipt = await waitToBeMined(this.provider, transaction.hash);
-      if (receipt.status) {
-        this.tokenContract.transfer(receipt.contractAddress, utils.parseEther('100'));
-      }
-    });
-
-    this.addKeySubscription = this.hooks.addListener('added', async (transaction) => {
-      const receipt = await waitToBeMined(this.provider, transaction.hash);
-      if (receipt.status) {
-        this.tokenContract.transfer(transaction.to, utils.parseEther('5'));
-      }
-    });
-
-    this.addKeysSubscription = this.hooks.addListener('keysAdded', async (transaction) => {
-      const receipt = await waitToBeMined(this.provider, transaction.hash);
-      if (receipt.status) {
-        this.tokenContract.transfer(transaction.to, utils.parseEther('15'));
-      }
-    });
-  }
-}
-
-export {TokenGrantingRelayer};
